Conflicts:
	Trigger/TrigAnalysis/TrigDecisionTool/Root/ExpertMethods.cxx
	Trigger/TrigAnalysis/TrigDecisionTool/Root/Logger.cxx
	Trigger/TrigAnalysis/TrigDecisionTool/Root/TrigDecisionTool.cxx
	Trigger/TrigAnalysis/TrigDecisionTool/Root/TrigDecisionToolCore.cxx
	Trigger/TrigAnalysis/TrigDecisionTool/TrigDecisionTool/Logger.h
	Trigger/TrigAnalysis/TrigDecisionTool/TrigDecisionTool/TrigDecisionTool.h
	Trigger/TrigAnalysis/TrigDecisionTool/src/DecisionUnpackerAthena.cxx

+++ Simulation/Tools/RDOAnalysis/src/TRT_FastRDOAnalysis.cxx
@@ -513,7 +513,7 @@ StatusCode TRT_FastRDOAnalysis::execute() {
         // Surface & Tracking info
         // -- element surface (straw layer)
         const Trk::Surface& surf((*detElement).surface());
-        const int surfType(surf.type());
+        const int surfType(static_cast<int>(surf.type()));
 
         const Trk::SurfaceBounds& bnds((*detElement).bounds());
         const int bndsType(bnds.type());
@@ -534,7 +534,7 @@ StatusCode TRT_FastRDOAnalysis::execute() {
 
         // -- straw surface (first straw in layer?)
         const Trk::Surface& strSurf((*detElement).surface(trtID));
-        const int strSurfType(strSurf.type());
+        const int strSurfType(static_cast<int>(strSurf.type()));
 
         const Trk::SurfaceBounds& strBnds((*detElement).bounds(trtID));
         const int strBndsType(strBnds.type());
